pyproject.toml
[project]
name = "ietp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "firebase-admin>=7.1.0",
    "jinja2>=3.1.6",
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
